backend_backup/requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
Flask-Cors>=3.0
Flask-Migrate>=4.0